add num_processing_errors column
